Refactor navigation and add back buttons to category screens
Introduced a `Routes.kt` object to centralize all navigation route constants, improving maintainability.

Refactored the home screen categories to use the new `Routes` for navigation to the `ArticlesScreen` and `AudioListScreen`.

Added a back arrow (`onNavigateBack`) to the `TopAppBar` in both the `ArticlesScreen` and `AudioListScreen` to allow users to navigate back to the previous screen.

Updated the UI theme and colors for `ArticlesScreen`, `AudioListScreen`, and `DetailArticleScreen` for better consistency.